Rename file to move_order_heuristic
