+++ sporkfish/searcher/searcher_config.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from sporkfish.searcher.move_ordering.move_ordering import MoveOrderMode
+from sporkfish.searcher.move_ordering.move_order_heuristic import MoveOrderMode
 
 from ..configurable import Configurable
